Implemented ControllerConfig in BasicDrive. Also, changed LTrack and RTrack to floats to work with ManualDrive.

+++ Avalanche/src/main/java/org/firstinspires/ftc/avalanche/utilities/ControllerConfig.java
@@ -13,8 +13,8 @@ import org.firstinspires.ftc.avalanche.subsystems.DriveTrainController;
 
 public interface ControllerConfig {
 
-    double LTrack();
-    double RTrack();
+    float LTrack();
+    float RTrack();
     boolean HarvesterButtonPressed();
     boolean ShooterButtonPressed();
     boolean LeftButtonPresserButtonPressed();
